Consolidate coordinator classes
